Restored d2 trans.in/out in working order

+++ src/libambulant/gui/d2/d2_img.cpp
@@ -40,8 +40,9 @@
 #endif
 
 using namespace ambulant;
+
 inline D2D1_RECT_F d2_rectf(lib::rect r) {
-	return D2D1::RectF(r.left(), r.top(), r.right(), r.bottom());
+	return D2D1::RectF((float) r.left(), (float) r.top(), (float) r.right(), (float) r.bottom());
 }
 
 extern const char d2_img_playable_tag[] = "img";
